+++ R/pat_createHourly.R
@@ -1,0 +1,119 @@
+#' @export
+#' @importFrom rlang .data
+#' @importFrom MazamaCoreUtils logger.error logger.debug logger.isInitialized
+#' @importFrom MazamaCoreUtils getAPIKey
+#'
+#' @title Create a new PurpleAir hourly timeseries dataset.
+#'
+#' @param api_key PurpleAir API Read Key. If \code{api_key = NULL}, it
+#' will be obtained using \code{getAPIKey("PurpleAir-read")}.
+#' See \code{MazamaCoreUtils::\link[MazamaCoreUtils:setAPIKey]{setAPIKey}}.
+#' @param pas Previously generated \emph{pas} object containing \code{sensor_index}.
+#' @param sensor_index PurpleAir sensor unique identifier.
+#' @param startdate Desired start time (ISO 8601) or \code{POSIXct}.
+#' @param enddate Desired end time (ISO 8601) or \code{POSIXct}.
+#' @param timezone Olson timezone used to interpret dates.
+#' @param fields Character string with PurpleAir field names for the Get Sensor Data API.
+#' @param baseUrl Base URL for the PurpleAir API.
+#' @param verbose Logical controlling the generation of warning and error messages.
+#'
+#' @return A PurpleAir Timeseries \emph{pat} object.
+#'
+#' @description Create a \code{pat} object for a specific \code{sensor_index}.
+#'
+#' By default, the only variables included in an "hourly pat" are:
+#' \code{datetime, humidity, temperature, pm2.5_atm, pm2.5_atm_a, pm2.5_atm_b}.
+#' These are sufficient for simple QC using \code{pm2.5_atm_a, pm2.5_atm_b} and
+#' simple correction using linear fit models based on
+#' \code{pm2.5_atm, temperature, humidity}.
+#'
+#' @references \href{https://www2.purpleair.com}{PurpleAir}
+#' @references \href{https://api.purpleair.com}{PurpleAir API}
+#' @references \href{https://www2.purpleair.com/policies/terms-of-service}{PurpleAir Terms of service}
+#' @references \href{https://www2.purpleair.com/pages/license}{PurpleAir Data license}
+#' @references \href{https://www2.purpleair.com/pages/attribution}{PurpleAir Data Attribution}
+#'
+#' @examples
+#' \donttest{
+#' # Fail gracefully if any resources are not available
+#' try({
+#'
+#' library(AirSensor2)
+#'
+#' pat <-
+#'   pat_createHourly(
+#'     api_key = PurpleAir_API_READ_KEY,
+#'     pas = MY_PAS,
+#'     sensor_index = "76545",
+#'     startdate = "2023-01-01",
+#'     enddate = "2023-01-08",
+#'     timezone = "UTC",
+#'     verbose = TRUE
+#'   )
+#'
+#' View(pat$meta[1:100,])
+#'
+#' }, silent = FALSE)
+#' }
+
+pat_createHourly <- function(
+    api_key = NULL,
+    pas = NULL,
+    sensor_index = NULL,
+    startdate = NULL,
+    enddate = NULL,
+    timezone = "UTC",
+    fields = PurpleAir_HISTORY_HOURLY_PM25_FIELDS,
+    baseUrl = "https://api.purpleair.com/v1/sensors",
+    verbose = FALSE
+) {
+
+  # ----- Validate parameters --------------------------------------------------
+
+  if ( is.null(api_key) )
+    api_key <- MazamaCoreUtils::getAPIKey("PurpleAir-read")
+
+  MazamaCoreUtils::stopIfNull(api_key)
+  MazamaCoreUtils::stopIfNull(pas)
+  MazamaCoreUtils::stopIfNull(sensor_index)
+  MazamaCoreUtils::stopIfNull(timezone)
+  MazamaCoreUtils::stopIfNull(fields)
+  MazamaCoreUtils::stopIfNull(baseUrl)
+  verbose <- MazamaCoreUtils::setIfNull(verbose, FALSE)
+
+  # ----- Create hourly pat ----------------------------------------------------
+
+  pat <- pat_createNew(
+    api_key = api_key,
+    pas = pas,
+    sensor_index = sensor_index,
+    startdate = startdate,
+    enddate = enddate,
+    timezone = timezone,
+    average = 60,
+    fields = fields,
+    baseUrl = baseUrl,
+    verbose = verbose
+  )
+
+  # ----- Return ---------------------------------------------------------------
+
+  return(pat)
+
+}
+
+# ===== DEBUGGING ==============================================================
+
+if ( FALSE ) {
+
+  api_key = PurpleAir_API_READ_KEY
+  pas = example_pas
+  sensor_index = "76545"
+  startdate = "2023-01-01"
+  enddate = "2023-01-08"
+  timezone = "America/Los_Angeles"
+  fields = PurpleAir_HISTORY_HOURLY_PM25_FIELDS
+  baseUrl = "https://api.purpleair.com/v1/sensors"
+  verbose = TRUE
+
+}